Exercise #2 Merged with our code and altered a bit
A lot was changed in terms of functions and functionality for the basic mechanics to work. Type "help" to see which commands we currently have and are usable. For the moment we can move only to Hangar #1 and back to Player Ship, as we will have to fix the naming of each room.